+++ src/main/java/subway/line/ui/LineController.java
@@ -6,7 +6,6 @@ import subway.line.application.LineService;
 import subway.line.application.dto.LineCreationDto;
 import subway.line.application.dto.StationAdditionToLineDto;
 import subway.line.application.dto.StationDeletionFromLineDto;
-import subway.line.ui.dto.DtoMapper;
 import subway.line.ui.dto.LineCreationRequest;
 import subway.line.ui.dto.LineDto;
 import subway.line.ui.dto.StationAdditionRequest;
